Added the full model for Financial Sanction Designations, as far as I am aware currently this is complete for the XML content but there may be more to add.
Full jaxb annotations in place for unmarshalling based on example xml
currently saved - at time of commit the online xml financial sanction
list content is not available, more probably to be updated once that is
in place.
Test in place for the unmarshalling but not complete. Does not yet check
assertions, details can be reviewed in debug for now.

Small additions for beginnings of online xml grabbing functionality.