Finished model, basic test, and javadocs for now.
Finished getting the model in place and documenting to some reasonable
extent. Changed test to have a more reasonable example xml and some
basic assertions, more to add in future.

The project should be approximately stable for front-end experimentation
at this point, although I will most likely stop for the moment until the
financial sanctions xml is back up on the gov site.